Move NameKind into Domain
