+++ src/Domain/QualifiedName.php
@@ -10,7 +10,7 @@ use Firehed\PhpLsp\Utility\NamespacePath;
  * A fully-qualified name split into namespace path and short name, carrying no
  * notion of what it names (Plan 0002 §5.3: the kind-neutral base FQN value type).
  * A name alone does not identify a symbol — PHP has three symbol namespaces — so a
- * {@see \Firehed\PhpLsp\Resolution\NameKind} travels alongside it.
+ * {@see NameKind} travels alongside it.
  *
  * The global namespace is the empty path.
  */
