+++ src/Knowledge/BuiltinBackend.php
@@ -8,11 +8,11 @@ use Firehed\PhpLsp\Domain\ClassInfo;
 use Firehed\PhpLsp\Domain\ClassName;
 use Firehed\PhpLsp\Domain\FunctionInfo;
 use Firehed\PhpLsp\Domain\FunctionName;
+use Firehed\PhpLsp\Domain\NameKind;
 use Firehed\PhpLsp\Domain\QualifiedName;
 use Firehed\PhpLsp\Index\NamespaceCatalog;
 use Firehed\PhpLsp\Index\NamespaceContents;
 use Firehed\PhpLsp\Repository\ClassInfoFactory;
-use Firehed\PhpLsp\Resolution\NameKind;
 use Psr\SimpleCache\CacheInterface;
 use ReflectionClass;
 use ReflectionException;
