+++ tests/Domain/NameKindTest.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace Firehed\PhpLsp\Tests\Resolution;
+namespace Firehed\PhpLsp\Tests\Domain;
 
+use Firehed\PhpLsp\Domain\NameKind;
 use Firehed\PhpLsp\Domain\QualifiedName;
-use Firehed\PhpLsp\Resolution\NameKind;
 use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\TestCase;
